Move EventStream to event::stream
