+++ examples/event-stream-tokio.rs
@@ -1,5 +1,5 @@
 //
-// cargo run --features async-event --example event-stream-tokio
+// cargo run --features event-stream --example event-stream-tokio
 //
 use std::io::{stdout, Write};
 use std::time::Duration;
